pyproject.toml
[project]
name = "ink" 
description = "Ink" 
readme = "README.md"
version = "1.0.0"
requires-python = ">=3.8"
license = {file = "LICENSE.md"}
dependencies = [
  "bardapi",
  "python-dotenv"
]
[build-system]
requires = ["setuptools>=43.0.0", "wheel"]
[tool.setuptools]
package-data = {"ink" = ["py.typed"]}
